feat(storage): make Parquet eager-decode ceiling operator-configurable (#473)
Expose the Parquet reader's eager-decode memory guard (design §5.4 C-DECODE) as an
internal ParquetDecodeLimits options record: the per-row-group decode ceiling (4 GiB
default) and the decompression-ratio ceiling (1000:1 default) are now operator-tunable
(lower below a constrained/multi-tenant executor budget, or raise for a trusted
large-row-group workload), with a validated fail-fast ctor and unchanged default
semantics. All new types are internal (no public API change).

RFL review also hardened the decompression-ratio check against 64-bit integer overflow:
`uncompressedBytes > compressedFloor * ratio` was an unchecked long*long multiply that
could wrap for a large declared compressed size, producing a wrong verdict. The
intermediate is now widened to Int128 (exact) at both guard sites that share the pattern
— ParquetFileReader.EnsureDecodeCeiling and DeltaCheckpointReader.ColumnFootprintBytes —
each pinned by a mutation-proven regression test.

Reviewed via the full multi-model RFL council (4 rounds) to unanimous 5/5 + decorrelated
Gemini red-team NO-MISS-CERTIFIED.

Closes #473

+++ src/DeltaSharp.Storage/Delta/DeltaCheckpointReader.cs
@@ -237,8 +237,9 @@ internal static class DeltaCheckpointReader
         }
 
         // Decompression-ratio ceiling — a chunk claiming far more decompressed than compressed bytes is a
-        // decompression bomb.
-        if (uncompressedBytes > Math.Max(compressedBytes, 1) * ParquetFileReader.MaxDecompressionRatio)
+        // decompression bomb. The product is widened to Int128 so a large declared compressed size cannot
+        // overflow the 64-bit multiply into a spurious verdict (wrapping past a bomb check).
+        if (uncompressedBytes > (Int128)Math.Max(compressedBytes, 1) * ParquetFileReader.MaxDecompressionRatio)
         {
             throw DeltaProtocolException.Malformed(string.Create(
                 CultureInfo.InvariantCulture,
